+++ eggs/Naaya/Naaya/CHANGELOG.rst
@@ -1,5 +1,6 @@
 4.0.14 (unreleased)
 -------------------
+* mediafile: skip encoding for compliant audio and mp4 max 720p [dumitval]
 * mediafile: show "still encoding" message, catch encodding error message [dumitval]
 * mediafile: fix for file upload on IE [dumitval]